LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/128-expected.txt

CONSOLE MESSAGE: line 2: SyntaxError: Unexpected end of script

Harness Error (FAIL), message = SyntaxError: Unexpected end of script

FAIL scheduler: appending script element to script  assert_array_equals: lengths differ, expected 5 got 3
